refactor(store): separate the similarity measure from its report
similarity.py was 715 lines: the measure (tokens, Jaccard,
containment, artifact corroboration, answer agreement, the verdict a
pair lands in) and the report that prints it, with the CLI at the
bottom. The measure moves to similarity_measure.py; similarity.py
keeps the git access, the false-cold candidates, the report, the
renderer and main, so the documented entry point is unchanged.

Under the limit either way — this is the file with the least headroom
left, and the two halves never needed each other's context.

ADVANCES pandoscope/agentic-engineering-template#239.
